Fixed categorization of Camunda
Camunda is a flexible enough to serve as a library as well as fully-fledged product. It is confusing to not have it in the fully-fledged product section. Also added Camunda Cloud as new SaaS offering and slightly adjusted Zeebe (the engine that fuels Camunda Cloud) wording to respect latest state of the project